monolithic fixes, benchmark update

+++ src/radix_gpu.cu
@@ -48,7 +48,7 @@ int main(int argc, char** argv) {
 
     // Benchmark example
     if (!conf.validation) {
-        bool ret = rsort::benchmark<false, uint8_t, size_t>(
+        bool ret = rsort::benchmark<false, uint32_t, size_t>(
             conf.n,
             conf.iterations,
             conf.warmups,
